Refactor out padding class from photo resubmission component
