+++ src/Patient/ReportingFlows/AltPhotoFlows/BottomButton.js
