+++ src/Patient/ReportingFlows/AltPhotoFlows/ContentContainer.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Box from '@material-ui/core/Box'
+
+const ContentContainer = (props) => {
+
+    return(<Box minHeight={"50vh"} padding="1em" {...props}  />);
+
+}
+
+export default ContentContainer;
