+++ src/Patient/index.js
@@ -17,7 +17,7 @@ import ForcePasswordChange from './ForcePasswordChange';
 import EducationalMessage from './Home/Education';
 import { usePageVisibility } from '../Hooks/PageVisibility';
 import UpdateContactTracing from './HouseholdTesting';
-import MissedPhotoFlow from './ReportingFlows/MissedPhotoFlow/';
+import MissedPhotoFlow from './ReportingFlows/AltPhotoFlows';
 import PushActionReportingFlow from './ReportingFlows/PushActionReportingFlow';
 import Box from '@material-ui/core/Box';
 
@@ -81,6 +81,11 @@ const PatientHome = observer(() => {
 
   if (uiStore.pathname.startsWith("/missed-photo")) {
     return <MissedPhotoFlow />
+  }
+
+
+  if (uiStore.pathname.startsWith("/redo-photo")) {
+    return <MissedPhotoFlow isRedo />
   }
 
   if (uiStore.pathname.startsWith("/quick-report")) {
